split module in 2
